Single project editor with BowTie.io support
* Introduces new authentication option for BowTie
* Removes user/repo routing requirement
* Removes repo browser
* Hardcode resource urls

This is part of a series of work that I'm performing to provide tools to
BowTie platform sites. This work is not suitable for inclusion upstream
since we've overwritten instead of conditionalized various routes and
features.

Includes `dist` since an archive will be referenced for use and we're
not requiring readers of the guide to perform a build.